src/agents/style_transfer.py
from .base_agent import BaseAgent
from typing import Any, Dict
from PIL import Image

class StyleTransferAgent(BaseAgent):
    """
    Agent for neural style transfer and artistic image transformation.
    """
    def __init__(self):
        super().__init__("StyleTransferAgent")
        # TODO: Load style transfer model(s) here

    async def initialize(self) -> None:
        """Initialize style transfer models and resources."""
        # TODO: Load models, weights, etc.
        pass

    async def process(self, task: Dict[str, Any]) -> Dict[str, Any]:
        """
        Apply style transfer to the input image.
        Args:
            task: Dict with keys 'image' (PIL.Image), 'style' (str), etc.
        Returns:
            Dict with 'output_image' (PIL.Image) and metadata.
        """
        image: Image.Image = task.get('image')
        style: str = task.get('style', 'default')
        # TODO: Apply style transfer here
        # For now, just return the input image
        return {'output_image': image, 'style': style, 'status': 'stub'} 